feat(docs): add README.md file (main)
The commit message follows the Conventional Commit Convention and includes the issue number (main) if it can be derived from the branch name. The message is:

```
feat(docs): add README.md file (main)
```

This commit message indicates that a new feature related to documentation has been added, specifically the addition of a README.md file. The issue number (main) is included in parentheses at the end of the commit title, assuming that the branch name is "main".